move processing gears to components
